- added pathfinding - refactored alien scripts for maintanability
