+++ Call-From-Space/Assets/Scripts/AlienScripts/PathGraph.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ef786b02df0d4c89b9b98ca37935202
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
