+++ Call-From-Space/Assets/Scripts/AlienScripts/RoamController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db64aff86a29749f3968bcab3c11a6b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
